+++ src/app/(protected)/[group]/[subGroup]/[instance]/supervisors/page.tsx
@@ -4,26 +4,16 @@ import { PageWrapper } from "@/components/page-wrapper";
 import { api } from "@/lib/trpc/server";
 import { InstanceParams } from "@/lib/validations/params";
 
-import { SupervisorsDataTable } from "./_components/supervisors-data-table";
+import { SupervisorsDataTable } from "./_components/all-supervisors-data-table";
 
 export default async function Page({ params }: { params: InstanceParams }) {
-  const user = await api.user.get();
   const role = await api.user.role({ params });
-
-  const stage = await api.institution.instance.currentStage({ params });
-  const tableData = await api.institution.instance.supervisors({
-    params,
-  });
+  const tableData = await api.institution.instance.supervisors({ params });
 
   return (
     <PageWrapper>
-      <Heading>Supervisors</Heading>
-      <SupervisorsDataTable
-        stage={stage}
-        user={user}
-        role={role}
-        data={tableData}
-      />
+      <Heading>All Supervisors</Heading>
+      <SupervisorsDataTable role={role} data={tableData} />
     </PageWrapper>
   );
 }
